fix: migrate breaking changes from application profile to v1 directory
