+++ src/leasing/v1/index.ts
@@ -1,2 +1,1 @@
 export * from './application-profile'
-export * as v1 from './v1'
