+++ src/schemas/v1/index.ts
@@ -1,4 +1,2 @@
 export * from './application-profile'
 export * from './application-profile-housing-reference'
-
-export * as v1 from './v1/'
